refactor(tests): split matrix_utils tests into domain files and add detection tests
migrate tests from the monolithic tests/test_matrix_utils.py to domain-specific files

delete the legacy tests/test_matrix_utils.py to prevent duplication

add tests/test_matrix_utils_detection.py to cover client/server capability detection

update tests/test_matrix_utils_core.py and tests/test_matrix_utils_error_handling.py to align with the split layout

update docs/dev/TESTING_GUIDE.md to reflect the new testing structure and note the removal of the monolith
